requirements.txt
boto==2.43.0
contextlib2==0.5.4
Django==1.9
django-admin-sortable2==0.6.4
django-ajax-selects==1.5.0
django-angular==0.7.15
django-autocomplete-light==3.1.8
django-extensions==1.5.9
django-pure-pagination==0.3.0
django-storages==1.5.1
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
easy-thumbnails==2.3
elasticsearch==5.0.1
fake-useragent==0.1.2
html5lib==1.0b8
lxml==3.6.4
oauth==1.0.1
Pillow==3.4.2
psycopg2==2.6.2
PyPDF2==1.26.0
rauth==0.7.2
raven==5.31.0
reportlab==3.3.0
requests==2.11.1
six==1.10.0
urllib3==1.19
webencodings==0.5
xhtml2pdf==0.0.6
